# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NovelHub",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/novels/{novel_id}/bookmarks": {
      "post": {
        "tags": [
          "bookmarks"
        ],
        "summary": "Create Novel Bookmark",
        "operationId": "create_novel_bookmark_api_novels__novel_id__bookmarks_post",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Novel Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookmarkCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "bookmarks"
        ],
        "summary": "List Novel Bookmarks",
        "operationId": "list_novel_bookmarks_api_novels__novel_id__bookmarks_get",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Novel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/novels/{novel_id}/bookmarks/{bookmark_id}": {
      "delete": {
        "tags": [
          "bookmarks"
        ],
        "summary": "Delete Novel Bookmark",
        "operationId": "delete_novel_bookmark_api_novels__novel_id__bookmarks__bookmark_id__delete",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Novel Id"
            }
          },
          {
            "name": "bookmark_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Bookmark Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete Novel Bookmark Api Novels  Novel Id  Bookmarks  Bookmark Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/novels/{novel_id}/progress": {
      "get": {
        "tags": [
          "bookmarks"
        ],
        "summary": "Get Novel Progress",
        "operationId": "get_novel_progress_api_novels__novel_id__progress_get",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Novel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "bookmarks"
        ],
        "summary": "Update Novel Progress",
        "operationId": "update_novel_progress_api_novels__novel_id__progress_put",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Novel Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgressResponse"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/analysis/start": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Start Analysis",
        "operationId": "start_analysis_api_analysis_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/analysis/{novel_id}/status": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Get Analysis Status",
        "operationId": "get_analysis_status_api_analysis__novel_id__status_get",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Novel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/analysis/{novel_id}/pause": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Pause Analysis",
        "operationId": "pause_analysis_api_analysis__novel_id__pause_post",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Novel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Pause Analysis Api Analysis  Novel Id  Pause Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/analysis/{novel_id}/resume": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Resume Analysis",
        "operationId": "resume_analysis_api_analysis__novel_id__resume_post",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Novel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Resume Analysis Api Analysis  Novel Id  Resume Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/analysis/{novel_id}/cancel": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Cancel Analysis",
        "operationId": "cancel_analysis_api_analysis__novel_id__cancel_post",
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Novel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Cancel Analysis Api Analysis  Novel Id  Cancel Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/backup/validate": {
      "post": {
        "tags": [
          "backup"
        ],
        "summary": "Validate Backup Endpoint",
        "operationId": "validate_backup_endpoint_api_backup_validate_post",
        "parameters": [
          {
            "name": "backup_data",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "contentMediaType": "application/octet-stream",
              "title": "Backup Data"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Validate Backup Endpoint Api Backup Validate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/backup/restore": {
      "post": {
        "tags": [
          "backup"
        ],
        "summary": "Restore Backup Endpoint",
        "operationId": "restore_backup_endpoint_api_backup_restore_post",
        "parameters": [
          {
            "name": "backup_data",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "contentMediaType": "application/octet-stream",
              "title": "Backup Data"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Restore Backup Endpoint Api Backup Restore Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalysisRequest": {
        "properties": {
          "novel_id": {
            "type": "integer",
            "title": "Novel Id"
          },
          "chapter_start": {
            "type": "integer",
            "title": "Chapter Start",
            "default": 1
          },
          "chapter_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter End"
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": "ollama"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": "llama3.2"
          }
        },
        "type": "object",
        "required": [
          "novel_id"
        ],
        "title": "AnalysisRequest"
      },
      "AnalysisStatus": {
        "properties": {
          "novel_id": {
            "type": "integer",
            "title": "Novel Id"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "chapters_processed": {
            "type": "integer",
            "title": "Chapters Processed"
          },
          "chapters_total": {
            "type": "integer",
            "title": "Chapters Total"
          },
          "entities_count": {
            "type": "integer",
            "title": "Entities Count"
          },
          "facts_count": {
            "type": "integer",
            "title": "Facts Count"
          },
          "errors": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "novel_id",
          "state",
          "chapters_processed",
          "chapters_total",
          "entities_count",
          "facts_count",
          "errors"
        ],
        "title": "AnalysisStatus"
      },
...
      "BookmarkCreate": {
        "properties": {
          "novel_id": {
            "type": "integer",
            "title": "Novel Id"
          },
          "chapter_id": {
            "type": "integer",
            "title": "Chapter Id"
          },
          "position": {
            "type": "integer",
            "title": "Position",
            "default": 0
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "novel_id",
          "chapter_id"
        ],
        "title": "BookmarkCreate"
      },
      "BookmarkList": {
        "properties": {
          "bookmarks": {
            "items": {
              "$ref": "#/components/schemas/BookmarkResponse"
            },
            "type": "array",
            "title": "Bookmarks"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "bookmarks",
          "total"
        ],
        "title": "BookmarkList"
      },
      "BookmarkResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "novel_id": {
            "type": "integer",
            "title": "Novel Id"
          },
          "chapter_id": {
            "type": "integer",
            "title": "Chapter Id"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "title": {
            "type": "string",
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "id",
          "novel_id",
          "chapter_id",
          "position",
          "title"
        ],
        "title": "BookmarkResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProgressResponse": {
        "properties": {
          "novel_id": {
            "type": "integer",
            "title": "Novel Id"
          },
          "chapter_id": {
            "type": "integer",
            "title": "Chapter Id"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          }
        },
        "type": "object",
        "required": [
          "novel_id",
          "chapter_id",
          "position"
        ],
        "title": "ProgressResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}